requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect==1.2.0
defusedxml==0.7.1
Django==3.2
django-admin-adminlte==1.0.5
djangorestframework==3.15.1
exceptiongroup==1.2.1
execnet==2.1.1
fasteners==0.19
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.2
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.0
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
pdbp==1.5.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
py==1.11.0
Pygments==2.17.2
pynose==1.5.1
pyotp==2.9.0
PySocks==1.7.1
pytest==8.2.0
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sbvirtualdisplay==1.3.0
selenium==4.20.0
selenium-stealth==1.0.6
seleniumbase==4.26.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tabcompleter==1.3.0
tomli==2.0.1
tornado==6.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
